refactor: event ui
