+++ code/cell.modules/pkg.sys/ui.primitives/src/Dev.Harness.tsx
@@ -15,9 +15,9 @@ const imports = {
   Card: import('./components/Card/DEV'),
   CardStack: import('./components/CardStack/dev/DEV'),
 
-  EventStack: import('./components/Event/dev/DEV.EventStack'),
-  EventCard: import('./components/Event/dev/DEV.EventCard'),
-  EventPipe: import('./components/Event/dev/DEV.EventPipe'),
+  EventCard: import('./components/Event.Card/DEV'),
+  EventStack: import('./components/Event.Stack/DEV'),
+  EventPipe: import('./components/Event.Pipe/DEV'),
 
   Button: import('./components.ref/button/Button.dev/DEV'),
   Switch: import('./components.ref/button/Switch.dev/DEV'),
