+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event.Card/EventCard.tsx
