+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event.Card/index.ts
@@ -1,0 +1,1 @@
+export * from './EventCard';
