+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event.Pipe/DEV.tsx
@@ -1,50 +1,40 @@
 import React from 'react';
 import { Observable, Subject } from 'rxjs';
-import { DevActions } from 'sys.ui.dev';
 
-import { EventStack, EventStackProps, useEventBusHistory } from '..';
-import { rx, t } from './common';
+import { DevActions } from 'sys.ui.dev';
+import { EventPipe, EventPipeProps } from '.';
+import { useEventBusHistory } from '../Event';
+import { t, rx } from '../../common';
 
 type Ctx = {
   bus: t.EventBus<any>;
   reset$: Subject<void>;
   count: number;
-  width?: number;
-  props: EventStackProps;
+  props: EventPipeProps;
 };
 
 /**
  * Actions
  */
 export const actions = DevActions<Ctx>()
-  .namespace('ui.event/EventStack')
+  .namespace('ui.event/EventPipe')
   .context((prev) => {
     if (prev) return prev;
 
     const bus = rx.bus();
     const reset$ = new Subject<void>();
 
-    const ctx: Ctx = {
+    return {
       bus,
       reset$,
       count: 0,
-      width: 300,
-      props: {},
+      props: { orientation: 'x' },
     };
-
-    return ctx;
   })
 
   .items((e) => {
     e.title('debug');
-
-    e.boolean('width', (e) => {
-      if (e.changing) e.ctx.width = e.changing.next ? 300 : undefined;
-      e.boolean.current = typeof e.ctx.width === 'number';
-    });
-
     e.button('reset', (e) => e.ctx.reset$.next());
-
     e.hr();
   })
 
@@ -53,15 +43,12 @@ export const actions = DevActions<Ctx>()
 
     e.select((config) => {
       config
-        .initial(config.ctx.props.card?.maxDepth || 3)
-        .title('card.maxDepth')
+        .initial(config.ctx.props.orientation)
+        .title('orientation')
         .view('buttons')
-        .items([0, 2, 3, 5, 10])
+        .items(['x', 'y'])
         .pipe((e) => {
-          if (e.changing) {
-            const props = e.ctx.props;
-            (props.card || (props.card = {})).maxDepth = e.changing.next[0].value;
-          }
+          if (e.changing) e.ctx.props.orientation = e.changing.next[0].value;
         });
     });
 
@@ -81,14 +68,17 @@ export const actions = DevActions<Ctx>()
   })
 
   .subject((e) => {
-    const { width, props, bus, reset$ } = e.ctx;
+    const { props, bus, reset$ } = e.ctx;
+
+    const size = 300;
+    const { orientation } = props;
+
     e.settings({
       host: { background: -0.04 },
       layout: {
-        label: { bottomLeft: '<EventStack>' },
         cropmarks: -0.2,
-        position: { top: 200 },
-        width,
+        width: orientation === 'x' ? size : undefined,
+        height: orientation === 'y' ? size : undefined,
       },
     });
     e.render(<Sample {...props} bus={bus} reset$={reset$} style={{ flex: 1 }} />);
@@ -96,9 +86,9 @@ export const actions = DevActions<Ctx>()
 
 export default actions;
 
-export type SampleProps = EventStackProps & { bus: t.EventBus<any>; reset$: Observable<void> };
+export type SampleProps = EventPipeProps & { bus: t.EventBus<any>; reset$: Observable<void> };
 export const Sample: React.FC<SampleProps> = (props) => {
   const { bus, reset$ } = props;
   const history = useEventBusHistory({ bus, reset$ });
-  return <EventStack {...props} events={history.events} style={{ flex: 1 }} />;
+  return <EventPipe {...props} events={history.events} style={{ flex: 1 }} />;
 };
