+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event.Pipe/index.ts
@@ -1,0 +1,1 @@
+export * from './EventPipe';
