+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event.Stack/DEV.tsx
@@ -1,39 +1,50 @@
 import React from 'react';
 import { Observable, Subject } from 'rxjs';
-
 import { DevActions } from 'sys.ui.dev';
-import { EventPipe, EventPipeProps, useEventBusHistory } from '..';
-import { t, rx } from './common';
+
+import { EventStack, EventStackProps, useEventBusHistory } from '.';
+import { rx, t } from '../../common';
 
 type Ctx = {
   bus: t.EventBus<any>;
   reset$: Subject<void>;
   count: number;
-  props: EventPipeProps;
+  width?: number;
+  props: EventStackProps;
 };
 
 /**
  * Actions
  */
 export const actions = DevActions<Ctx>()
-  .namespace('ui.event/EventPipe')
+  .namespace('ui.event/EventStack')
   .context((prev) => {
     if (prev) return prev;
 
     const bus = rx.bus();
     const reset$ = new Subject<void>();
 
-    return {
+    const ctx: Ctx = {
       bus,
       reset$,
       count: 0,
-      props: { orientation: 'x' },
+      width: 300,
+      props: {},
     };
+
+    return ctx;
   })
 
   .items((e) => {
     e.title('debug');
+
+    e.boolean('width', (e) => {
+      if (e.changing) e.ctx.width = e.changing.next ? 300 : undefined;
+      e.boolean.current = typeof e.ctx.width === 'number';
+    });
+
     e.button('reset', (e) => e.ctx.reset$.next());
+
     e.hr();
   })
 
@@ -42,12 +53,15 @@ export const actions = DevActions<Ctx>()
 
     e.select((config) => {
       config
-        .initial(config.ctx.props.orientation)
-        .title('orientation')
+        .initial(config.ctx.props.card?.maxDepth || 3)
+        .title('card.maxDepth')
         .view('buttons')
-        .items(['x', 'y'])
+        .items([0, 2, 3, 5, 10])
         .pipe((e) => {
-          if (e.changing) e.ctx.props.orientation = e.changing.next[0].value;
+          if (e.changing) {
+            const props = e.ctx.props;
+            (props.card || (props.card = {})).maxDepth = e.changing.next[0].value;
+          }
         });
     });
 
@@ -67,17 +81,14 @@ export const actions = DevActions<Ctx>()
   })
 
   .subject((e) => {
-    const { props, bus, reset$ } = e.ctx;
-
-    const size = 300;
-    const { orientation } = props;
-
+    const { width, props, bus, reset$ } = e.ctx;
     e.settings({
       host: { background: -0.04 },
       layout: {
+        label: { bottomLeft: '<EventStack>' },
         cropmarks: -0.2,
-        width: orientation === 'x' ? size : undefined,
-        height: orientation === 'y' ? size : undefined,
+        position: { top: 200 },
+        width,
       },
     });
     e.render(<Sample {...props} bus={bus} reset$={reset$} style={{ flex: 1 }} />);
@@ -85,9 +96,9 @@ export const actions = DevActions<Ctx>()
 
 export default actions;
 
-export type SampleProps = EventPipeProps & { bus: t.EventBus<any>; reset$: Observable<void> };
+export type SampleProps = EventStackProps & { bus: t.EventBus<any>; reset$: Observable<void> };
 export const Sample: React.FC<SampleProps> = (props) => {
   const { bus, reset$ } = props;
   const history = useEventBusHistory({ bus, reset$ });
-  return <EventPipe {...props} events={history.events} style={{ flex: 1 }} />;
+  return <EventStack {...props} events={history.events} style={{ flex: 1 }} />;
 };
