+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event.Stack/EventStack.tsx
@@ -2,8 +2,9 @@ import React, { useRef, useState } from 'react';
 
 import { css, CssValue, defaultValue, useResizeObserver } from '../../common';
 import { CardStack, CardStackItem } from '../CardStack';
-import { cardFactory as defaultCardFactory } from './factory';
-import { EventStackCardFactory, EventLogItem } from './types';
+import { cardFactory as defaultCardFactory } from './EventStack.factory';
+import { EventLogItem } from '../Event/types';
+import { EventStackCardFactory } from './types';
 
 export type EventStackProps = {
   events?: EventLogItem[];
