+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event.Stack/index.ts
@@ -1,0 +1,2 @@
+export * from './EventStack';
+export { useEventBusHistory } from '../Event';
