+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event.Stack/types.ts
@@ -1,0 +1,9 @@
+import { EventLogItem } from '../Event/types';
+
+export type EventStackCardFactory = (args: EventStackCardFactoryArgs) => JSX.Element;
+export type EventStackCardFactoryArgs = EventLogItem & {
+  width: number;
+  isTopCard: boolean;
+  showPayload?: boolean;
+  toggleShowPayload?: () => void;
+};
