+++ code/cell.modules/pkg.sys/ui.primitives/src/components/Event/index.ts
@@ -1,6 +1,2 @@
-export * from './EventStack';
-export * from './EventCard';
-export * from './EventPipe';
-
 export * from './types';
 export * from './useEventBusHistory';
